Don't divide by zero when rounding in invoice.
It could happen that a duration is rounded to zero for an invoice.
In that case a DivisionByZero exception was thrown.

The offending line tries to calculate a rate so that this rate
multiiplied by th
e rounded hours gives the amount of the entry.
We just leave the rate as it is, if the rounded hour is zero.

closes #267