Fixed starting multiple recordings. (fixes #419)
It was possible to start multiple recordings by applying a filter so
that the current recording was not shown anymore. This worked because
the running recording was determined from the list of shown entries.

The fix is to always check in the database if a recording is running and
return its information.