Japanese Language Localisation: (jp.php). [0.9.3]
This is the Japanese Language Localisation file for Kimai Version 0.9.3 (Based of the en.php file from the master branch of kimai).


NOTES:

In Japanese, Task is easier to understand than Activity thus the translation is of task. Should you for whatever reason wish for it strongly to read activity change タスク to 活動. This will make grammatical sense but may be somewhat difficult to understand in terms of parts of a project.

The lines of code 253-256 will not make sense the way they are.
253	"roundTimesheetEntries" => "新タイムシートエントリーを "<time to round>" に丸める",  
254	"minutes" => "分",
255	"seconds" => "秒",
256	"and" => "と",	
It needs to be arranged that roughly translated from Japanese to English it reads 「New Timesheet Entry <Mintues> <and> <Seconds> to round to.」 To try and explain using the lines of code the order would be [Half of 253]>[254]>[256]>[255]>[Half of 253].
In Japanese it should read 新タイムシートエントリーを　#分　と　#秒　に丸める

The line of code 264 still shows 'show_sensible_data’ as a language variable.
264	'show_sensible_data' => 'デバッグエクステンションにセンシブルデータを表示する',
As the comment in my last merge request and https://github.com/kimai/kimai/issues/375 says, I assume this will be deleted or hidden.

The lines of code 253-256 will not make sense the way they are.
293	'editLimitPart1' => '日間',
294	'editLimitPart2' => '時間',
295	'editLimitPart3' => '以前に終わるエントリーは作成と編集をできません',
It needs to be arranged that roughly translated from Japanese to English {ie. It won’t be grammatical English} it reads「<Days> <Hours> ending before in the past can not be added or edited.」 In other words, the order can stay the same but the day variable needs to go first in the string (before 293) and the time variable needs to go before line 294 followed by line 295.
In Japanese it should read #日間　#時間　 以前に終わるエントリーは作成と編集をできません

The line of code 414 will appear blank.
414	6 => "",
As per the previous comments on my last merge request, this line is not required in Japanese as it does not make grammatical sense to write anything before the subject <backup date>. The "Backup from" line has been added to the following line and this line is left blank.
Arranging this string roughly translated from Japanese to English would produce 「<backup date> backup has been recovered.」

If anything needs more explaining or changing then please don't hesitate to ask me.