Update en.php
As per the discussion in Pull Request #454. (Same file with the removal of ?>).


Some spelling and grammar edits.

If there is something out of place please let me know. Also it is assumed this in American English. (Otherwise Cell would be Mobile etc.).

There is also a non consecutive use of 「 ' 」 and 「 " 」 if it means anything in particular (I am not sure exactly in which file each item is called so this is probably nothing).

There is room to add capital letters to certain field names for aesthetic purposes as well.