Recordings can always be stopped. (fixes #455)
Up until now the stopRecord call always used an ID of 0 to identify the
timesheet entry to stop. The database layer had a special logic for this
case to use the latest (most recent created) timesheet entry. If for any
reason another entry was created for that user this entry selected
instead of the one we wanted to stop. The entry was actually not changed
because an additional check ensures that only running entries are
stopped by stopRecord.

I also fixed highlighting the correct row to indicate that a recording
is being stopped.