Don't leak project names and activities in exports.
A previous fix (f950fe92863942060ec0dfd7061a0219a89f90d0) fetched all
customers, projects and activities from the database for the groups of
the current user. For customers this did not work and returned a list of
all customers, projects and activities.

For the list of customers this was not an issue since it's later set to
only be the logged in customer.

For the list of activities this was no problem either because this list
was intersected with the selected activities and then only used for
filtering.

For the list of projects this was a problem. It was used to display the
names of the selected projects in several export formats.

This fix now determines the list of projects and activities the customer
may actually see first.

Fixes: #473