[MIGRATION,FEATURE] tinyDoc-->openTBS; support for .docx templates
Added openTBS support for supporting MS Word invoice templates. Due to
the changes the naming of the blocks in the templates had to be changed.
Changed generation of 'invoiceId' in ki_invoice/print.php (german
regulation needs increasing invoice numbers).
Invoice filename is generated starting with invoiceId.
Added support for numbering positions of a invoice with tag
[pos.position] in invoice templates (see template 2018)
All .odt templates must be migrated:
1. replace all [var.XXXX] tags to [kimai.XXXX]
2. replace all [row.XXXX] tags to [pos.XXXX]
Use of .doc(x) templates is similar to .odt.

All odt templates are migrated to openTBS.