Update en.php to show "City, State" in floater for United States (#1329)
* Fix issue #1327
* Update en.php

This is just a simple change that I made so that it is clear to users in the United States to input City, State together (e.g. Fort Myers, FL) in the City field when adding new customers.  On the Add Customer floater, it now requests "City, State:" instead of just "City:" 

Then, when you set up your invoice templates, it pulls city & state together so that the mailing address is properly formatted.